+++ tibiawikisql/models/npc.py
@@ -1,16 +1,14 @@
-import contextlib
 import sqlite3
 from sqlite3 import Connection, Cursor
 from typing import Any
 
 from pydantic import BaseModel, Field
-from pypika import Parameter, SQLLiteQuery as Query, Table
 from typing_extensions import Self
 
 from tibiawikisql.api import WikiEntry
 from tibiawikisql.models.base import RowModel, WithImage, WithStatus, WithVersion
 from tibiawikisql.schema import NpcBuyingTable, NpcDestinationTable, NpcJobTable, NpcRaceTable, NpcSellingTable, \
-    NpcSpellTable, NpcTable
+    NpcTable
 
 
 class NpcOffer(BaseModel):
@@ -27,78 +25,6 @@ class NpcOffer(BaseModel):
     value: int
     """The value of the item being offered."""
 
-
-
-class TaughtSpell(BaseModel):
-    """A spell taught by an NPC."""
-    spell_title: str
-    """The title of the article containing the spell's details."""
-    spell_id: int
-    """The article ID of the spell."""
-    knight: bool
-    """If the spell is taught to knights."""
-    paladin: bool
-    """If the spell is taught to paladins."""
-    druid: bool
-    """If the spell is taught to druids."""
-    sorcerer: bool
-    """If the spell is taught to sorcerers."""
-    monk: bool
-    """If the spell is taught to monks."""
-    price: int
-
-class NpcSpell(RowModel, table=NpcSpellTable):
-    """Represents a spell that a NPC can teach."""
-
-    npc_id: int
-    """The article id of the npc that teaches the spell."""
-    spell_id: int
-    """The article id of the spell taught by the npc."""
-    knight: bool
-    """If the spell is taught to knights."""
-    paladin: bool
-    """If the spell is taught to paladins."""
-    druid: bool
-    """If the spell is taught to druids."""
-    sorcerer: bool
-    """If the spell is taught to sorcerers."""
-    monk: bool
-    """If the spell is taught to monks."""
-
-    def insert(self, conn: sqlite3.Connection | sqlite3.Cursor) -> None:
-        if self.spell_id is not None:
-            super().insert(conn)
-
-        spell_table = Table(NpcSpellTable.__tablename__)
-        npc_spell_table = Table(self.table.__tablename__)
-        q = (
-            Query.into(npc_spell_table)
-            .columns(
-                "npc_id",
-                "spell_id",
-                "knight",
-                "paladin",
-                "druid",
-                "sorcerer",
-                "monk",
-            )
-            .insert(
-                Parameter(":npc_id"),
-                (
-                    Query.from_(spell_table)
-                    .select(spell_table.article_id)
-                    .where(spell_table.title == Parameter(":spell_title"))
-                ),
-                Parameter(":knight"),
-                Parameter(":paladin"),
-                Parameter(":druid"),
-                Parameter(":sorcerer"),
-                Parameter(":monk"),
-            )
-        )
-        query_str = q.get_sql()
-        with contextlib.suppress(sqlite3.IntegrityError):
-            conn.execute(query_str, self.model_dump(mode="json"))
 
 
 class NpcDestination(BaseModel):
@@ -158,8 +84,6 @@ class Npc(WikiEntry, WithVersion, WithStatus, WithImage, RowModel, table=NpcTabl
     """Items bought by the NPC."""
     destinations: list[NpcDestination] = Field(default_factory=list)
     """Places where the NPC can travel to."""
-    teaches: list[TaughtSpell] = Field(default_factory=list)
-    """Spells this NPC can teach."""
 
     @property
     def job(self) -> str | None:
@@ -194,18 +118,6 @@ class Npc(WikiEntry, WithVersion, WithStatus, WithImage, RowModel, table=NpcTabl
             return None
         npc.jobs = [j["name"] for j in NpcJobTable.get_list_by_field(conn, "npc_id", npc.article_id)]
         npc.races = [j["name"] for j in NpcRaceTable.get_list_by_field(conn, "npc_id", npc.article_id)]
-        npc.teaches = [
-            TaughtSpell(
-                spell_title=r["title"],
-                spell_id=r["article_id"],
-                price=r["price"],
-                knight=r["knight"],
-                paladin=r["paladin"],
-                sorcerer=r["sorcerer"],
-                druid=r["druid"],
-                monk=r["monk"],
-            ) for r in NpcSpellTable.get_by_npc_id(conn, npc.article_id)
-        ]
         npc.sell_offers = [
             NpcOffer(
                 item_id=r["item_id"],